Updating with latest test project and bugfixes.

+++ KoreCommon/Image/KoreSkiaSharpConv.cs
@@ -72,11 +72,13 @@ public static class KoreSkiaSharpConv
             (byte)(alpha * 255));
     }
 
+    // Usage: KoreColorRGB kCol = KoreSkiaSharpConv.ToKoreColorRGB(skCol);
+
     public static KoreColorRGB ToKoreColorRGB(this SKColor color)
     {
-        float colR = (float)color.Red;
-        float colG = (float)color.Green;
-        float colB = (float)color.Blue;
+        byte colR = (byte)color.Red;
+        byte colG = (byte)color.Green;
+        byte colB = (byte)color.Blue;
 
         return new KoreColorRGB(colR, colG, colB);
     }
